extracted functionality to composable
