+++ frontend/components/Domain/QueryFilterBuilder.vue
@@ -269,56 +269,10 @@ import draggable from "vuedraggable";
 import { computed, defineComponent, reactive, ref, toRefs, watch } from "@nuxtjs/composition-api";
 import { useHouseholdSelf } from "~/composables/use-households";
 import RecipeOrganizerSelector from "~/components/Domain/Recipe/RecipeOrganizerSelector.vue";
-import { Organizer, RecipeOrganizer } from "~/lib/api/types/non-generated";
+import { Organizer } from "~/lib/api/types/non-generated";
 import { LogicalOperator, QueryFilterJSON, QueryFilterJSONPart, RelationalKeyword, RelationalOperator } from "~/lib/api/types/response";
 import { useCategoryStore, useFoodStore, useHouseholdStore, useTagStore, useToolStore } from "~/composables/store";
-import { FieldLogicalOperator, FieldRelationalOperator, useQueryFilterBuilder } from "~/composables/use-query-filter-builder";
-
-interface OrganizerBase {
-  id: string | number;
-  name: string;
-}
-
-export type FieldType =
-  | "string"
-  | "number"
-  | "boolean"
-  | "date"
-  | RecipeOrganizer;
-
-export type FieldValue =
-  | string
-  | number
-  | boolean
-  | Date
-  | Organizer;
-
-interface SelectableItem {
-  label: string;
-  value: FieldValue;
-};
-
-export interface FieldDefinition {
-  name: string;
-  label: string;
-  type: FieldType;
-
-  // only for select/organizer fields
-  fieldOptions?: SelectableItem[];
-}
-
-interface Field extends FieldDefinition {
-  leftParenthesis?: string;
-  logicalOperator?: FieldLogicalOperator;
-  value: FieldValue;
-  relationalOperatorValue: FieldRelationalOperator;
-  relationalOperatorOptions: FieldRelationalOperator[];
-  rightParenthesis?: string;
-
-  // only for select/organizer fields
-  values: FieldValue[];
-  organizers: OrganizerBase[];
-}
+import { Field, FieldDefinition, FieldValue, OrganizerBase, useQueryFilterBuilder } from "~/composables/use-query-filter-builder";
 
 export default defineComponent({
   components: {
@@ -337,7 +291,7 @@ export default defineComponent({
   },
   setup(props, context) {
     const { household } = useHouseholdSelf();
-    const { logOps, relOps } = useQueryFilterBuilder();
+    const { logOps, relOps, buildQueryFilterString, getFieldFromFieldDef, isOrganizerType } = useQueryFilterBuilder();
 
     const firstDayOfWeek = computed(() => {
       return household.value?.preferences?.firstDayOfWeek || 0;
@@ -358,16 +312,6 @@ export default defineComponent({
       [Organizer.Household]: useHouseholdStore(),
     };
 
-    function isOrganizerType(type: FieldType): type is Organizer {
-      return (
-        type === Organizer.Category ||
-        type === Organizer.Tag ||
-        type === Organizer.Tool ||
-        type === Organizer.Food ||
-        type === Organizer.Household
-      );
-    }
-
     function onDragEnd(event: any) {
       state.drag = false;
 
@@ -379,71 +323,6 @@ export default defineComponent({
       const field = fields.value.splice(oldIndex, 1)[0];
       fields.value.splice(newIndex, 0, field);
     }
-
-    function getFieldFromFieldDef(field: Field | FieldDefinition, resetValue = false): Field {
-      const updatedField = {logicalOperator: logOps.value.AND, ...field} as Field;
-      let operatorOptions: FieldRelationalOperator[];
-      if (updatedField.fieldOptions?.length || isOrganizerType(updatedField.type)) {
-        operatorOptions = [
-          relOps.value["IN"],
-          relOps.value["NOT IN"],
-          relOps.value["CONTAINS ALL"],
-        ];
-      } else {
-        switch (updatedField.type) {
-          case "string":
-            operatorOptions = [
-              relOps.value["="],
-              relOps.value["<>"],
-              relOps.value["LIKE"],
-              relOps.value["NOT LIKE"],
-            ];
-            break;
-          case "number":
-            operatorOptions = [
-              relOps.value["="],
-              relOps.value["<>"],
-              relOps.value[">"],
-              relOps.value[">="],
-              relOps.value["<"],
-              relOps.value["<="],
-            ];
-            break;
-          case "boolean":
-            operatorOptions = [relOps.value["="]];
-            break;
-          case "date":
-            operatorOptions = [
-            relOps.value["="],
-              relOps.value["<>"],
-              relOps.value[">"],
-              relOps.value[">="],
-              relOps.value["<"],
-              relOps.value["<="],
-            ];
-            break;
-          default:
-            operatorOptions = [relOps.value["="], relOps.value["<>"]];
-        }
-      }
-      updatedField.relationalOperatorOptions = operatorOptions;
-      if (!operatorOptions.includes(updatedField.relationalOperatorValue)) {
-        updatedField.relationalOperatorValue = operatorOptions[0];
-      }
-
-      if (resetValue) {
-        updatedField.value = "";
-        updatedField.values = [];
-        updatedField.organizers = [];
-      } else {
-        updatedField.value = updatedField.value || "";
-        updatedField.values = updatedField.values || [];
-        updatedField.organizers = updatedField.organizers || [];
-      }
-
-
-      return updatedField;
-    };
 
     const fields = ref<Field[]>([]);
 
@@ -526,81 +405,6 @@ export default defineComponent({
       state.datePickers.splice(index, 1);
     };
 
-    function buildQueryFilterString() {
-      let isValid = true;
-      let lParenCounter = 0;
-      let rParenCounter = 0;
-
-      const parts: string[] = [];
-      fields.value.forEach((field, index) => {
-        if (index) {
-          if (!field.logicalOperator) {
-            field.logicalOperator = logOps.value.AND;
-          }
-          parts.push(field.logicalOperator.value);
-        }
-
-        if (field.leftParenthesis && state.showAdvanced) {
-          lParenCounter += field.leftParenthesis.length;
-          parts.push(field.leftParenthesis);
-        }
-
-        if (field.label) {
-          parts.push(field.name);
-        } else {
-          isValid = false;
-        }
-
-        if (field.relationalOperatorValue) {
-          parts.push(field.relationalOperatorValue.value);
-        } else {
-          if (field.type !== "boolean") {
-            isValid = false;
-          }
-        }
-
-        if (field.fieldOptions?.length || isOrganizerType(field.type)) {
-          if (field.values?.length) {
-            let val: string;
-            if (field.type === "string" || field.type === "date" || isOrganizerType(field.type)) {
-              val = field.values.map((value) => `"${value}"`).join(",");
-            } else {
-              val = field.values.join(",");
-            }
-            parts.push(`[${val}]`);
-          } else {
-            isValid = false;
-          }
-        } else {
-          if (field.value) {
-            if (field.type === "string" || field.type === "date") {
-              parts.push(`"${field.value}"`);
-            } else {
-              parts.push(field.value.toString());
-            }
-          } else {
-            if (field.type === "boolean") {
-              parts.push("false");
-            } else {
-              isValid = false;
-            }
-          }
-        }
-
-        if (field.rightParenthesis && state.showAdvanced) {
-          rParenCounter += field.rightParenthesis.length;
-          parts.push(field.rightParenthesis);
-        }
-      });
-
-      if (lParenCounter !== rParenCounter) {
-        isValid = false;
-      }
-
-      state.qfValid = isValid;
-      return isValid ? parts.join(" ") : "";
-    }
-
     watch(
       // Toggling showAdvanced changes the builder logic without changing the field values,
       // so we need to manually trigger reactivity to re-run the builder.
@@ -626,10 +430,11 @@ export default defineComponent({
           fields.value[index] = updatedField;
         });
 
-        const qf = buildQueryFilterString();
+        const qf = buildQueryFilterString(fields.value, state.showAdvanced);
         if (qf) {
           console.debug(`Set query filter: ${qf}`);
         }
+        state.qfValid = !!qf;
 
         context.emit("input", qf || undefined);
       },
@@ -680,7 +485,7 @@ export default defineComponent({
           return initFieldsError(`"Invalid query filter; unknown attribute name "${part.attributeName}"`);
         }
 
-        const field: Field = getFieldFromFieldDef(fieldDef);
+        const field = getFieldFromFieldDef(fieldDef);
         field.leftParenthesis = part.leftParenthesis || field.leftParenthesis;
         field.rightParenthesis = part.rightParenthesis || field.rightParenthesis;
         field.logicalOperator = part.logicalOperator ?
@@ -789,7 +594,7 @@ export default defineComponent({
         },
       }
 
-      return attrs
+      return attrs;
     })
 
     return {
